pyproject.toml
[tool.poetry]
name = "myweb"
version = "0.1.0"
description = ""
authors = ["Alex Wang <alexwang0405@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
